# docs/swagger_api_gateway_auth.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Query Pipeline — API Gateway",
        "description": "Orchestrates the full NL→MongoDB→LLM pipeline. Accepts a natural language query, converts it to a MongoDB query via query-service, executes it, and returns a report or video prompt generated by llm-service.",
        "contact": {
            "name": "Pipeline Team",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login with username and password",
                "description": "Returns a JWT bearer token valid for the configured duration. The token is stored in MongoDB and invalidated automatically when it expires. Pass it as `Authorization: Bearer <token>` on all protected endpoints.",
                "operationId": "login_api_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pipeline": {
            "post": {
                "tags": [
                    "Pipeline"
                ],
                "summary": "Execute end-to-end pipeline",
                "description": "Receives a natural language input, orchestrates query-service to produce a MongoDB query, executes that query, then calls llm-service to generate either a structured markdown report or a video generation prompt.",
                "operationId": "run_pipeline_api_pipeline_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PipelineRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PipelineResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid mode or request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDetail"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Upstream service failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDetail"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Downstream service error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDetail"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Downstream service unreachable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDetail"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/pipeline/{pipeline_id}": {
            "get": {
                "tags": [
                    "Pipeline"
                ],
                "summary": "Retrieve pipeline result",
                "description": "Fetch a previously executed pipeline result from the MongoDB audit collection.",
                "operationId": "get_pipeline_api_pipeline__pipeline_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "pipeline_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Pipeline Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Pipeline Api Pipeline  Pipeline Id  Get"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pipeline ID not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDetail"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDetail"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorDetail": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "ErrorDetail"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LoginRequest": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password"
                ],
                "title": "LoginRequest",
                "examples": [
                    {
                        "password": "yourpassword",
                        "username": "admin"
                    }
                ]
            },
            "MongoQueryResult": {
                "properties": {
                    "collection": {
                        "type": "string",
                        "title": "Collection"
                    },
                    "filter": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Filter"
                    },
                    "projection": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Projection"
                    },
                    "sort": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sort"
                    },
                    "limit": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Limit"
                    }
                },
                "type": "object",
                "required": [
                    "collection",
                    "filter"
                ],
                "title": "MongoQueryResult"
            },
            "PipelineRequest": {
                "properties": {
                    "user_input": {
                        "type": "string",
                        "title": "User Input",
                        "description": "Natural language request"
                    },
                    "mode": {
                        "type": "string",
                        "enum": [
                            "report",
                            "video"
                        ],
                        "title": "Mode",
                        "description": "Output mode"
                    },
                    "collection_hint": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Collection Hint",
                        "description": "Suggested collection name"
                    },
                    "context": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Context",
                        "description": "Extra context for LLM"
                    }
                },
                "type": "object",
                "required": [
                    "user_input",
                    "mode"
                ],
                "title": "PipelineRequest",
                "examples": [
                    {
                        "collection_hint": "instagram_posts",
                        "context": "Concentrati sui livelli di coinvolgimento",
                        "mode": "report",
                        "user_input": "Seleziona i post del 18-5-2026 con almeno 5 mi piace"
                    }
                ]
            },
            "PipelineResponse": {
                "properties": {
                    "pipeline_id": {
                        "type": "string",
                        "title": "Pipeline Id"
                    },
                    "mode": {
                        "type": "string",
                        "enum": [
                            "report",
                            "video"
                        ],
                        "title": "Mode"
                    },
                    "mongo_query": {
                        "$ref": "#/components/schemas/MongoQueryResult"
                    },
                    "raw_data": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Raw Data"
                    },
                    "result": {
                        "type": "string",
                        "title": "Result"
                    },
                    "tokens_used": {
                        "type": "integer",
                        "title": "Tokens Used"
                    },
                    "duration_ms": {
                        "type": "integer",
                        "title": "Duration Ms"
                    }
                },
                "type": "object",
                "required": [
                    "mode",
                    "mongo_query",
                    "raw_data",
                    "result",
                    "tokens_used",
                    "duration_ms"
                ],
                "title": "PipelineResponse",
                "examples": [
                    {
                        "duration_ms": 1423,
                        "mode": "report",
                        "mongo_query": {
                            "collection": "instagram_posts",
                            "filter": {
                                "likesCount": {
                                    "$gte": 5
                                },
                                "timestamp": {
                                    "$gte": "2026-05-18T00:00:00.000Z",
                                    "$lte": "2026-05-18T23:59:59.000Z"
                                }
                            },
                            "limit": 100,
                            "sort": {
                                "likesCount": -1
                            }
                        },
                        "pipeline_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                        "raw_data": [
                            {
                                "caption": "Lamborghini Urus knows no compromise...",
                                "commentsCount": 0,
                                "embedded_images": [
                                    {
                                        "file_size_bytes": 84210,
                                        "filename": "CfUqpZgqnPv_1.jpg",
                                        "image_index": 1
                                    }
                                ],
                                "hashtags": [
                                    "LuxuryCarRental",
                                    "sardegna"
                                ],
                                "likesCount": 47,
                                "ownerUsername": "luxury_car_division",
                                "post_id": "CfUqpZgqnPv",
                                "timestamp": "2026-05-18T13:36:10.000Z",
                                "type": "Sidecar"
                            }
                        ],
                        "result": "## Sommario\n\nSono stati trovati 12 post pubblicati il 18 maggio 2026 con almeno 5 mi piace...",
                        "tokens_used": 512
                    }
                ]
            },
            "TokenResponse": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "default": "bearer"
                    },
                    "expires_in_minutes": {
                        "type": "integer",
                        "title": "Expires In Minutes"
                    }
                },
                "type": "object",
                "required": [
                    "access_token",
                    "expires_in_minutes"
                ],
                "title": "TokenResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}